requirements.txt
numpy>=2.0,<2.6
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
pyyaml>=6.0
xgboost>=2.0
lightgbm>=4.0
implicit>=0.7
faiss-cpu>=1.8
torch>=2.0
matplotlib>=3.8
pytest>=8.0
psutil>=5.9
